Consolidate UI modules into terminal-ui and update import paths
