+++ src/cli/prompts/prompt-manager.ts
@@ -1,6 +1,6 @@
 import { readFileSync, existsSync } from 'fs';
 import { join } from 'path';
-import { CliUI } from '../utils/cli-ui';
+import { CliUI } from '../ui/terminal-ui';
 
 /**
  * PromptManager - Sistema per gestire e caricare system prompts specifici
